+++ webpack.config.js
@@ -5,18 +5,32 @@ import path from 'path';
 import { fileURLToPath } from 'url';
 const __dirname = fileURLToPath(import.meta.url);
 
-let config = {
+const srcPath = path.resolve(__dirname, '../src');
+const entryPath = './docs/assets/js/src';
+const distPath = path.resolve(__dirname, '../docs/assets/js/dist');
+
+const config = {
 	mode: process.env.MODE,
-	entry: './docs/assets/js/src/main.js',
+	entry: {
+		'docs-script': `${entryPath}/docs-script.ts`,
+	},
 	output: {
-		path: path.resolve(__dirname, 'docs/assets/js/dist'),
-		filename: 'bundle.js',
+		path: distPath,
+		filename: '[name].bundle.js',
 	},
 	resolve: {
+		extensions: ['.js', '.ts'],
 		alias: {
-			csv: path.resolve(__dirname, '../csv'),
+			'@cipscis/csv': `${srcPath}/csv.ts`,
 		},
-		extensions: ['.js'],
+	},
+	module: {
+		rules: [
+			{
+				test: /\.ts$/,
+				loader: 'ts-loader',
+			},
+		],
 	},
 };